# api/api/openapi.yml
openapi: '3.0.3'
info:
  title: AusCERT 2023 - Going back to the basics of API security
  description: API documentation for a demo API.
  contact:
    name: Kyle Jackson
    email: kyle.jackson@forgeresponse.com
  version: 1.0.0
servers:
  - url: https://auscert.forgeresponse.com
components:
  schemas:
...
    Webhook:
      type: object
      properties:
        id:
          type: string
        algorithm:
          type: string
        description:
          type: string
        events:
          type: array
          items:
            type: string
        secret:
          type: string
        url:
          type: string
    CreateWebhookRequest:
      type: object
      properties:
        description:
          type: string
        events:
          type: array
          items:
            type: string
        url:
          type: string
    UpdateWebhookRequest:
      type: object
      properties:
        description:
          type: string
        events:
          type: array
          items:
            type: string
        url:
          type: string
    ErrorResponse:
      type: object
      properties:
        id:
          type: string
        message:
          type: string
        code:
          type: integer
        transactionId:
          type: string
  securitySchemes:
    ApiKeyAuth:
      type: apiKey
      name: Authorization
      in: header
security:
  - ApiKeyAuth: []
paths:
...
  /v1/webhooks:
    get:
      description: Returns a list of webhooks
      tags:
        - Webhooks
      parameters:
        - in: query
          name: limit
          schema:
            type: integer
          description: Specify the number of items to return
      responses:
        '200':
          description: Successfully returned a list of webhooks
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Webhook'
        '404':
          description: No records found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
    post:
      description: Creates a new webhook
      tags:
        - Webhooks
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateWebhookRequest'
      responses:
        '200':
          description: Successfully created webhook
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Webhook'
        '400':
          description: Invalid request body
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /v1/webhooks/{id}:
    parameters:
      - name: id
        in: path
        description: Id of the user
        required: true
        schema:
          type: string
    get:
      description: Returns the specified webhook
      tags:
        - Webhooks
      responses:
        '200':
          description: Successfully returned webhook
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Webhook'
        '404':
          description: No record found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
    put:
      description: Updates the specified webhook
      tags:
        - Webhooks
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateWebhookRequest'
      responses:
        '200':
          description: Successfully updated webhook
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Webhook'
        '400':
          description: Invalid request body
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

    delete:
      description: Deletes the specified webhook
      tags:
        - Webhooks
      responses:
        '204':
          description: Successfully deleted webhook
        '404':
          description: No record found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /v1/webhooks/{id}/rotate:
    parameters:
      - name: id
        in: path
        description: Id of the user
        required: true
        schema:
          type: string
    post:
      description: Rotates the secret for the specified webhook
      tags:
        - Webhooks
      responses:
        '200':
          description: Successfully rotated webhook secret
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Webhook'
        '400':
          description: Invalid request body
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'